sched: Introduce Per-Cluster DVFS Headroom
Typically, UI-demanding tasks are handled by little and big cores,
while the prime core is more power-hungry and leads to unnecessary boosts,
causing power wastage. From my observations, boosting the little cores
provides the most significant performance improvement for UI tasks,
followed by a modest boost to the big cores. This results in fewer
stutters across the entire system.

To avoid excessive boosting, the implementation ensures that when
the utilization is equal to or higher than the maximum capacity value,
util is returned as-is.